MdePkg/Include: Update SevSnpMsr.h and Msr.h header guard defines
Update the header guards in SevSnpMsr.h and Msr.h to match filename  and to
comply with latest edk2 coding guidelines.  This change is part of a refactor
of MSR definitions and SEV-SNP related defines, which aims to remove
family-specific references (filename) as these defines are common to all
modern EPYC Processors.

Signed-off-by: Paul Grimes <paul.grimes@amd.com>